Add writethroughFile_darwin (although the writethrough behaviour is not yet actually implemented)
