+++ common/mmf_darwin.go
@@ -24,7 +24,6 @@ package common
 
 import (
 	"os"
-	"strings"
 	"sync"
 	"syscall"
 )
@@ -89,37 +88,3 @@ func (m *MMF) Slice() []byte {
 	return m.slice
 }
 
-// create a file, given its path and length
-func CreateFileOfSize(destinationPath string, fileSize int64) (*os.File, error) {
-	CreateParentDirectoryIfNotExist(destinationPath)
-
-	f, err := os.OpenFile(destinationPath, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0644)
-	if err != nil {
-		return nil, err
-	}
-	//err = syscall.Fallocate(int(f.Fd()), 0, 0, fileSize)
-	//if err != nil {
-	//	return nil, err
-	//}
-	// TODO: Need to appropriate fallocate api for darwin
-	if truncateError := f.Truncate(fileSize); truncateError != nil {
-		return nil, truncateError
-	}
-	return f, nil
-}
-
-func CreateParentDirectoryIfNotExist(destinationPath string) error {
-	// check if parent directory exists
-	parentDirectory := destinationPath[:strings.LastIndex(destinationPath, AZCOPY_PATH_SEPARATOR_STRING)]
-	_, err := os.Stat(parentDirectory)
-	// if the parent directory does not exist, create it and all its parents
-	if os.IsNotExist(err) {
-		err = os.MkdirAll(parentDirectory, os.ModePerm)
-		if err != nil {
-			return err
-		}
-	} else if err != nil {
-		return err
-	}
-	return nil
-}
